requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip==0.2.0
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.79.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
python-docx==1.1.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sympy==1.14.0
tbb==2021.13.1
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
upstash-vector==0.8.0
urllib3==2.4.0
Werkzeug==3.1.3
